Make mmappedfile that's read/write use a writable mmap
Currently a read/write mmappedfile uses a readonly mmapped region for
reading, but then write() syscalls to do any writes into the file.

This appears to cause sub-optimal interactions with the linux kernel. We
see when doing lots of twoskip work (e.g. a conversations db recosntruct
on a large user) a lot of kernel time like:

  Overhead  Shared Object            Symbol
    22.99%  [kernel]                 [k] filemap_map_pages
    19.88%  [kernel]                 [k] unmap_page_range
     7.04%  libcyrus.so.0.0.0        [.] read_onerecord
     3.89%  libcyrus.so.0.0.0        [.] crc32
     3.26%  [kernel]                 [k] xas_load
     2.00%  [kernel]                 [k] asm_exc_page_fault
     1.78%  [kernel]                 [k] clear_page_rep
     1.70%  [kernel]                 [k] __mod_node_page_state

This patch attempts to provide an alternative approach. When opening a
mmappedfile as read/write, we actually map it twice. Once as a readonly
mapping which we use for all read accesses, but a secondary read/write
mapping. We keep these separate as a basic safety precaution to reduce
the chance of accidental corruptions.

So then when we want to write to the file, rather than calling write()
syscalls, we just memcpy() into the writable mapping.

However the underlying kernel page mapping for both regions should be
to the same pages, just with different permissions. So this should avoid
any kernel unmap/map actions and save a lot of kernel time when doing
any writes to mmappedfile files (e.g. two skip databases).